added CheckoutIntercom on CheckoutPayment
