+++ src/client/pages/Checkout/CheckoutPayment/CheckoutPayment.tsx
@@ -4,6 +4,7 @@ import { useQuoteCartData } from 'utils/hooks/useQuoteCartData'
 import { CheckoutPageWrapper } from '../shared/CheckoutPageWrapper'
 import { Footer } from '../shared/Footer'
 import { PaymentInfo } from '../shared/PaymentInfo'
+import { CheckoutIntercomVariation } from '../shared/CheckoutIntercomVariation'
 import { ContactInformation } from './ContactInformation/ContactInformation'
 import { PaymentDetails } from './PaymentDetails/PaymentDetails'
 
@@ -20,6 +21,7 @@ export const CheckoutPayment = () => {
     <CheckoutPageWrapper>
       <ContactInformation {...userDetails} />
       <PaymentDetails />
+      <CheckoutIntercomVariation />
       <Footer
         buttonText={textKeys.CHECKOUT_FOOTER_CONTINUE_TO_PAYMENT()}
         buttonOnClick={() => {
